+++ worker/acl_cache_test.go
@@ -10,7 +10,7 @@
  *     https://github.com/dgraph-io/dgraph/blob/master/licenses/DCL.txt
  */
 
-package edgraph
+package worker
 
 import (
 	"encoding/json"
@@ -21,14 +21,10 @@ import (
 )
 
 func TestAclCache(t *testing.T) {
-	aclCachePtr = &aclCache{
-		predPerms: make(map[string]map[string]int32),
-	}
-
 	var emptyGroups []string
 	group := "dev"
 	predicate := "friend"
-	require.Error(t, aclCachePtr.authorizePredicate(emptyGroups, predicate, acl.Read),
+	require.Error(t, AuthorizePredicate(emptyGroups, predicate, acl.Read),
 		"the anonymous user should not have access when the acl cache is empty")
 
 	acls := []acl.Acl{
@@ -44,16 +40,16 @@ func TestAclCache(t *testing.T) {
 			Acls:    string(aclBytes),
 		},
 	}
-	aclCachePtr.update(groups)
+	Update(groups)
 	// after a rule is defined, the anonymous user should no longer have access
-	require.Error(t, aclCachePtr.authorizePredicate(emptyGroups, predicate, acl.Read),
+	require.Error(t, AuthorizePredicate(emptyGroups, predicate, acl.Read),
 		"the anonymous user should not have access when the predicate has acl defined")
-	require.NoError(t, aclCachePtr.authorizePredicate([]string{group}, predicate, acl.Read),
+	require.NoError(t, AuthorizePredicate([]string{group}, predicate, acl.Read),
 		"the user with group authorized should have access")
 
 	// update the cache with empty acl list in order to clear the cache
-	aclCachePtr.update([]acl.Group{})
+	Update([]acl.Group{})
 	// the anonymous user should have access again
-	require.Error(t, aclCachePtr.authorizePredicate(emptyGroups, predicate, acl.Read),
+	require.Error(t, AuthorizePredicate(emptyGroups, predicate, acl.Read),
 		"the anonymous user should not have access when the acl cache is empty")
 }
